requirements.txt
anthropic
gradio
python-dotenv
elevenLabs